refactor: move resource entity/form files to proper subdirectories
