+++ amd/src/resource_checklist/resource_checklist_category.js
@@ -34,7 +34,7 @@ import ResourceChecklistItem from './resource_checklist_item';
 const CATEGORY_REGION = 'resource-checklist-category';
 const ITEM_REGION = 'resource-checklist-item-row';
 const CATEGORY_MODAL_FORM = 'mod_bookit\\form\\edit_category_form';
-const ITEM_MODAL_FORM = 'mod_bookit\\form\\edit_resource_checklist_item_form';
+const ITEM_MODAL_FORM = 'mod_bookit\\local\\form\\resource\\edit_resource_checklist_item_form';
 
 /**
  * Resource checklist category component.
