+++ classes/local/entity/resource/bookit_resource_checklist.php
@@ -23,7 +23,7 @@
  * @license     https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace mod_bookit\local\entity;
+namespace mod_bookit\local\entity\resource;
 
 /**
  * Entity class for resource checklist.
