+++ classes/local/install_helper.php
@@ -28,8 +28,8 @@ namespace mod_bookit\local;
 use mod_bookit\local\entity\masterchecklist\bookit_checklist_master;
 use mod_bookit\local\entity\masterchecklist\bookit_checklist_category;
 use mod_bookit\local\entity\masterchecklist\bookit_checklist_item;
-use mod_bookit\local\entity\bookit_resource_category;
-use mod_bookit\local\entity\bookit_resource;
+use mod_bookit\local\entity\resource\bookit_resource_category;
+use mod_bookit\local\entity\resource\bookit_resource;
 use mod_bookit\local\manager\resource_manager;
 
 /**
