+++ classes/local/manager/resource_manager.php
@@ -24,8 +24,8 @@
 namespace mod_bookit\local\manager;
 
 use dml_exception;
-use mod_bookit\local\entity\bookit_resource;
-use mod_bookit\local\entity\bookit_resource_category;
+use mod_bookit\local\entity\resource\bookit_resource;
+use mod_bookit\local\entity\resource\bookit_resource_category;
 use mod_bookit\local\manager\resource_checklist_manager;
 
 /**
