+++ classes/output/resource_category_card.php
@@ -25,7 +25,7 @@
 
 namespace mod_bookit\output;
 
-use mod_bookit\local\entity\bookit_resource_category;
+use mod_bookit\local\entity\resource\bookit_resource_category;
 use mod_bookit\local\manager\resource_manager;
 use renderer_base;
 use renderable;
